Add pixeltable/tests/functions subfolder and refactor
