+++ pixeltable/tests/functions/test_fireworks.py
