+++ pixeltable/tests/functions/test_openai.py
